AL-2128: Add compiler flags Wshadow and Wextra, and apply fixes for issues identified

+++ ST/hw_rtc.c
@@ -320,7 +320,7 @@ TimerTime_t HW_RTC_Tick2ms( uint32_t tick )
 void HW_RTC_SetAlarm( uint32_t timeout )
 {
   /* we don't go in Low Power mode for timeout below MIN_ALARM_DELAY */
-  if ( (MIN_ALARM_DELAY + McuWakeUpTimeCal ) < ((timeout - HW_RTC_GetTimerElapsedTime( ) )) )
+  if ( (uint32_t) (MIN_ALARM_DELAY + McuWakeUpTimeCal ) < ((timeout - HW_RTC_GetTimerElapsedTime( ) )) )
   {
     LowPower_Enable( e_LOW_POWER_RTC );
   }
